Mount /buildbot/git readonly for tools that need .git symlinks to be valid.
When building on a keystone build machine $GITDIR entries are symlinks that
point somewhere in a tree of git repos rooted at /buildbot/git. We mount this
directory so that tools that rely on those symlinks being valid will not fail.
We mount readonly to prevent the build from manipulating source control
information. This directory is not present on local machines, so we don't
require this mount to present.

Also remove the workaround for b/119317944

Bug: 119317944
Test: Build succeeds on both local machine and android build machine, python nsjail_test.py
Change-Id: I94c190e78ee7add2f66a3308680741ba92286c5b